backend/requirements.txt
python-dotenv
requests
openai
fastapi
uvicorn
pydantic
gunicorn